% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrnr_alan_pois.R
\docType{class}
\name{Lrnr_alan_pois}
\alias{Lrnr_alan_pois}
\title{Generalized Linear Models}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_alan_pois
}
\value{
Learner object with methods for training and prediction. See
 \code{\link{Lrnr_base}} for documentation on learners.
}
\description{
This learner provides fitting procedures for generalized linear models using
\code{\link[stats]{glm.fit}}.
}
\section{Parameters}{

\describe{
  \item{\code{...}}{Parameters passed to \code{\link[stats]{glm}}.}
}
}

\seealso{
Other Learners: \code{\link{Lrnr_SIR}},
  \code{\link{Lrnr_stratify_horizon}},
  \code{\link{Lrnr_transform_outcome}}
}
\concept{Learners}
\keyword{data}
