% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_SIR.R
\docType{class}
\name{Lrnr_SIR}
\alias{Lrnr_SIR}
\title{SIR}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_SIR
}
\value{
Learner object with methods for training and prediction. See
 \code{\link{Lrnr_base}} for documentation on learners.
}
\description{
This learner solves a simple SIR model
}
\section{Parameters}{

\describe{
  \item{\code{...}}{Parameters passed to \code{\link[stats]{glm}}.}
}
}

\seealso{
Other Learners: \code{\link{Lrnr_alan_pois}},
  \code{\link{Lrnr_stratify_horizon}},
  \code{\link{Lrnr_transform_outcome}}
}
\concept{Learners}
\keyword{data}
