% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_transform_outcome.R
\docType{class}
\name{Lrnr_transform_outcome}
\alias{Lrnr_transform_outcome}
\title{Bound Predictions}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_transform_outcome
}
\value{
Learner object with methods for training and prediction. See
 \code{\link{Lrnr_base}} for documentation on learners.
}
\description{
This learner bounds predictions. Intended for use in a pipeline.
}
\section{Parameters}{

\describe{
  \item{\code{bound = .005}}{Either a length two vector of c(lower,upper) or a
  lower bound, where the upper is then 1 - lower}
  \item{\code{...}}{Not currently used.}
}
}

\seealso{
Other Learners: \code{\link{Lrnr_SIR}},
  \code{\link{Lrnr_alan_pois}},
  \code{\link{Lrnr_stratify_horizon}}
}
\concept{Learners}
\keyword{data}
